Simplify logic, combine tests, add 1% test
D6 override
